Issue #1344: added PocketViewModel skeleton
